separate GameControl from actions and simplify the logic
