+++ src/game_control.rs
@@ -1,0 +1,35 @@
+use bevy::prelude::{Input, KeyCode, Res};
+
+pub enum GameControl {
+    Up,
+    Down,
+    Left,
+    Right,
+}
+
+impl GameControl {
+    pub fn pressed(&self, keyboard_input: &Res<Input<KeyCode>>) -> bool {
+        match self {
+            GameControl::Up => {
+                keyboard_input.pressed(KeyCode::W) || keyboard_input.pressed(KeyCode::Up)
+            }
+            GameControl::Down => {
+                keyboard_input.pressed(KeyCode::S) || keyboard_input.pressed(KeyCode::Down)
+            }
+            GameControl::Left => {
+                keyboard_input.pressed(KeyCode::A) || keyboard_input.pressed(KeyCode::Left)
+            }
+            GameControl::Right => {
+                keyboard_input.pressed(KeyCode::D) || keyboard_input.pressed(KeyCode::Right)
+            }
+        }
+    }
+}
+
+pub fn is_pressed(control: GameControl, input: &Res<Input<KeyCode>>) -> f32 {
+    if control.pressed(input) {
+        1.0
+    } else {
+        0.0
+    }
+}
