+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod actions;
 mod audio;
+mod game_control;
 mod loading;
 mod menu;
 mod player;
